Tried analyzing packages, added new flag to tsdumper, added 32bit dump to bituitls, updated some PAS stuff.

+++ ts.h
@@ -4,6 +4,6 @@
 
 void MPEG_transport_stream(char* filename);
 void transport_packet(char* data);
-int program_association_section(char* data, uint8_t pos);
+uint8_t program_association_section(char* data, uint8_t pos);
 uint8_t adaptation_field(char* data);
 #endif
